requirements.txt
astroid==2.3.3
bcrypt==3.1.7
boto3==1.12.45
botocore==1.15.45
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
hvac==0.10.0
idna==2.8
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
py4j==0.10.7
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
pysftp==0.2.8
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
